requirements.txt
pandas
google-cloud-bigquery
google-api-python-client
google-auth
flask
openpyxl
matplotlib
fpdf
